+++ src/router/index.ts
@@ -3,13 +3,13 @@ import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router'
 const Demo = () => import('@/views/Demo.vue')
 const page404 = () => import('@/views/404.vue')
 const Home = () => import('@/views/home/home.vue')
-const clickPage = () => import('@/views/clickPage/clickPage.vue')
+// const clickPage = () => import('@/views/clickPage/clickPage.vue')
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
-    name: 'clickPage',
-    component: clickPage,
+    name: '/',
+    component: Home,
     meta: {
       title: ''
     }
