+++ src/views/home/components/clickPage.vue
